+++ llamazure/azgraph/azgraph.py
@@ -5,7 +5,7 @@ import dataclasses
 import json
 import operator
 from functools import reduce
-from typing import Any, Tuple, Union, cast
+from typing import Any, Tuple, cast
 
 import requests
 
@@ -63,11 +63,11 @@ class Graph:
 		).json()
 		return cast(Tuple[str], tuple(s["subscriptionId"] for s in raw["value"]))
 
-	def q(self, q: str) -> Union[Any, ResErr]:
+	def q(self, q: str) -> Any:
 		"""Make a graph query"""
 		res = self.query(Req(q, self.subscriptions))
 		if isinstance(res, ResErr):
-			return res
+			raise res.exception()
 		return res.data
 
 	def _exec_query(self, req) -> ResMaybe: