+++ llamazure/tools/migrate/dashboard.py
@@ -8,10 +8,9 @@ import click
 from azure.identity import DefaultAzureCredential
 
 from llamazure.azrest.azrest import AzRest
-from llamazure.azrest.models import cast_as
 from llamazure.rid import rid
 from llamazure.rid.rid import Resource
-from llamazure.tools.migrate.portal.r.m.portal.portal import AzDashboards, Dashboard, PatchableDashboard  # pylint: disable=E0611,E0401
+from llamazure.tools.migrate.portal.r.m.portal.portal import AzDashboards  # pylint: disable=E0611,E0401
 from llamazure.tools.migrate.util import JSONTraverser, rid_params
 
 
@@ -43,10 +42,11 @@ class Migrator:
 
 	def put_dashboard(self, transformed: dict):
 		"""Update the dashboard in Azure with the transformed data."""
-		d = Dashboard(**transformed)
-		p = cast_as(d, PatchableDashboard)
+		# they broke the openapi model lol
+		# d = Dashboard(**transformed)
+		# p = cast_as(d, PatchableDashboard)
 		self.az.call(
-			AzDashboards.Update(*rid_params(self.dashboard), p),
+			AzDashboards.Update(*rid_params(self.dashboard), transformed),  # type: ignore
 		)
 
 	def make_backup(self, dashboard: dict):
@@ -71,9 +71,10 @@ class Restorer:
 			return json.load(f)
 
 	def put_dashboard(self, content: dict):
-		p = PatchableDashboard(**content)
+		# they broke the openapi model lol
+		# p = PatchableDashboard(**content)
 		self.az.call(
-			AzDashboards.Update(*rid_params(self.dashboard), p),
+			AzDashboards.Update(*rid_params(self.dashboard), content),  # type: ignore
 		)