+++ llamazure/msgraph/changelog.md
@@ -1,5 +1,12 @@
 # 0
 
+## 0.1
+
+### 0.1.0
+
+- feature: **breaking** `q` throws on error
+- task: fix namespace package
+
 # 0.0
 
 ### 0.0.3