+++ llamazure/tools/changelog.md
@@ -2,6 +2,11 @@
 
 ## 0.0
 
+### 0.2.0
+
+- task: fix namespace package
+- task: workaround for broken dashboard openapi model
+
 ### 0.1.0
 
 - feature: migrate workspace of dashboards